requirements.txt
openai
python-dotenv
backoff